apps/recruitment/selenium-automation/requirements.txt
browser-use>=0.7.9
python-dotenv>=1.0.0
playwright>=1.40.0
selenium>=4.15.0
